Makes copy operations async
Copy operations blocking the main thread is an increasing problem as
filesizes get larger and/or the connection gets slower.  This aims to
solve the former problem and prep for changes to alleviate the latter.

Solving the former:
  The cycle-times involved in copying large files start to block the
main thread fairly quickly. This should allow the main loop to keep
running even in the face of many changes.

Partial help for the latter:
  As the connection gets slower, files may also be generated faster
than they are transferred out. This has both the effect of the former
(large files) and a much higher probability of causing multiple copy
operations to happen simultaneously. This change, though, should allow
for an application copy queue to be built to alleviate the second
problem.