Bugfixes - config loading
- Adds loading of example_project_config.json - lots of these values are
required and not loading this file leads to missing them in the
configuration and having blind issues. Better to handle it here and have
the program fail early than have weird troubleshooting problems later.
- Fixes 'projecft' typo
- Fixes config.config_options vs. raw config object issue in
ScipionTemplate class
- Changes minimum_import_interval in default configuration file to null.
This should not be preset by the install but set by the administrator on
configuration.